fix: move pluginDagSort to universal folder
