+++ packages/builder/builder-rspack-provider/src/core/initPlugins.ts
@@ -80,7 +80,9 @@ export async function initPlugins({
 }) {
   debug('init plugins');
 
-  const { pluginDagSort } = await import('@modern-js/utils');
+  const { pluginDagSort } = await import(
+    '@modern-js/utils/universal/pluginDagSort'
+  );
 
   const plugins = pluginDagSort(pluginStore.plugins);
