+++ packages/builder/builder-webpack-provider/src/core/initPlugins.ts
@@ -81,7 +81,9 @@ export async function initPlugins({
 }) {
   debug('init plugins');
 
-  const { pluginDagSort } = await import('@modern-js/utils');
+  const { pluginDagSort } = await import(
+    '@modern-js/utils/universal/pluginDagSort'
+  );
 
   const plugins = pluginDagSort(pluginStore.plugins);
