+++ packages/toolkit/utils/src/index.ts
@@ -42,4 +42,3 @@ export * from './getCoreJsVersion';
 export * from './react';
 export * from './getTargetDir';
 export * from './babel';
-export * from './pluginDagSort';
